Added confirmation request before deletion to prevent accidental removal

+++ build/k8s-delete.sh
@@ -27,4 +27,15 @@ if [ ! -d "${instance}" ]; then
   exit 1
 fi
 
-oc delete -f ${instance}/target/k8s/namespace.yml
+echo "Warning: you are about to delete ${instance}. This may result in data loss."
+echo "To confirm that you know what you're doing, please type in the path to the"
+echo "instance (relative to current directory) you would like to remove."
+
+read -p "Instance path: " confirm
+
+if [[ "$(readlink -f ${confirm})" == "$(readlink -f ${instance})" ]]; then
+  echo "Deleting ${instance}..."
+  oc delete -f ${instance}/target/k8s/namespace.yml
+else
+  echo "Cannot delete, $(readlink -f ${confirm}) != $(readlink -f ${instance})"
+fi
